requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Camera discovery and ONVIF
requests==2.31.0
lxml==4.9.3
zeep==4.2.1
onvif-zeep==0.2.12

# Face recognition (optional for development)
# Uncomment these for actual face recognition functionality
# opencv-python==4.8.1.78
# face-recognition==1.3.0
# numpy==1.24.3
# Pillow==10.1.0

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1